requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
typing-extensions==3.10.0.0
